Add support for Chip tooltip
