+++ src/panel_material_ui/pane/Chip.jsx
@@ -6,7 +6,9 @@ const SIZES = {
   medium: "2em",
 }
 
-export function render({model}) {
+export function render(props, ref) {
+  const {model, ...other} = props
+
   const [color] = model.useState("color")
   const [disabled] = model.useState("disabled")
   const [icon] = model.useState("icon")
@@ -15,16 +17,23 @@ export function render({model}) {
   const [variant] = model.useState("variant")
   const [sx] = model.useState("sx")
 
+  if (Object.entries(ref).length === 0 && ref.constructor === Object) {
+    ref = React.useRef(null)
+  }
+
   return (
     <Chip
       color={color}
       disabled={disabled}
+      fullWidth
       icon={icon ? render_icon(icon, null, size, SIZES[size]) : null}
       label={render_icon_text(label)}
+      onClick={(e) => model.send_event("click", e)}
+      ref={ref}
       size={size}
       sx={sx}
       variant={variant}
-      onClick={(e) => model.send_event("click", e)}
+      {...other}
     />
   )
 }
