+++ src/panel_material_ui/pane/base.py
@@ -7,7 +7,7 @@ import param
 from panel.links import Callback
 from panel.pane import HoloViews, Markdown
 
-from ..base import COLORS, MaterialComponent
+from ..base import COLORS, MaterialComponent, ThemedTransform, TooltipTransform
 from ..widgets import DatetimeInput, DiscreteSlider, EditableFloatSlider, EditableIntSlider, FloatSlider, IntSlider, Select
 
 HoloViews.default_widgets = dict(
@@ -195,7 +195,7 @@ class Avatar(ClickablePaneBase):
     _esm_base = "Avatar.jsx"
 
 
-class Chip(ClickablePaneBase):
+class Chip(ClickablePaneBase, TooltipTransform):
     """
     A `Chip` can be used to display information, labels, tags, or actions. It can include text,
     an avatar, an icon, or a delete button.
@@ -215,6 +215,12 @@ class Chip(ClickablePaneBase):
         Color theme of the chip. Available options include standard Material UI colors
         like 'primary', 'secondary', 'default', 'error', 'info', 'success', 'warning'."""
     )
+
+    description = param.String(default="", doc="Tooltip text to display when hovering over the widget.")
+
+    description_delay = param.Integer(default=500, doc="""
+        Delay (in milliseconds) to display the tooltip after the cursor has
+        hovered over the Button, default is 500ms.""")
 
     disabled = param.Boolean(
         default=False,
@@ -253,6 +259,8 @@ class Chip(ClickablePaneBase):
     )
 
     _esm_base = "Chip.jsx"
+    _esm_transforms = [TooltipTransform, ThemedTransform]
+
 
 
 class Skeleton(MaterialPaneBase):
